requirements.txt
numpy==1.26.4
streamlit==1.32.0
google-generativeai==0.8.4
shap==0.46.0
matplotlib==3.8.4
tensorflow==2.18.0